requirements.txt
# Async HTTP
aiohttp>=3.9,<4.0

# Конфигурация
python-dotenv>=1.0
PyYAML>=6.0

# База данных
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13

# Redis
redis>=5.0

# Логирование
loguru>=0.7

# Telegram
aiogram>=3.4

# Веб-дашборд (Фаза 5)
fastapi>=0.109
uvicorn>=0.27
jinja2>=3.1

# PDF отчёты (Фаза 4)
reportlab>=4.1

# Экспорт (Фаза 5)
openpyxl>=3.1

# Тесты
pytest>=8.0
pytest-asyncio>=0.23
aioresponses>=0.7
